# api/openapi.yaml
openapi: 3.0.3
info:
  title: CS Inventory Edit Local Agent API
  version: 0.1.0
  description: >-
    Contract for communication between the application shells and the local
    agent. Endpoints are added here as they migrate from the legacy handwritten
    contract.
servers:
  - url: http://127.0.0.1:7331
paths:
...
  /games/{game}/inventory:
    parameters:
      - $ref: "#/components/parameters/Game"
    get:
      operationId: getGameInventory
      responses:
        "200": { $ref: "#/components/responses/GameInventoryResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
  /games/{game}/inventory/refresh:
    parameters:
      - $ref: "#/components/parameters/Game"
    post:
      operationId: refreshGameInventory
      responses:
        "200": { $ref: "#/components/responses/OperationReceiptResponse" }
        "403": { $ref: "#/components/responses/OperationReceiptResponse" }
...
  /steam-inventory-service/{appID}:
    parameters:
      - $ref: "#/components/parameters/AppID"
    get:
      operationId: getSteamInventoryService
      responses:
        "200": { $ref: "#/components/responses/GameInventoryResponse" }
        "400": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
  /steam-inventory-service/{appID}/refresh:
    parameters:
      - $ref: "#/components/parameters/AppID"
    post:
      operationId: refreshSteamInventoryService
      responses:
        "200": { $ref: "#/components/responses/OperationReceiptResponse" }
        "400": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/OperationReceiptResponse" }
  /armory/redeem:
    post:
      operationId: redeemArmory
      requestBody: { $ref: "#/components/requestBodies/ArmoryRedeem" }
      responses:
        "200": { $ref: "#/components/responses/OperationReceiptResponse" }
...
  /games/tf2/store/purchases:
    post:
      operationId: initializeTf2StorePurchase
      requestBody:
        { $ref: "#/components/requestBodies/InitializeStorePurchase" }
      responses:
        "200": { $ref: "#/components/responses/PurchaseSessionResponse" }
        "400": { $ref: "#/components/responses/ErrorResponse" }
  /store/purchases:
    post:
      operationId: initializeStorePurchase
      requestBody:
        { $ref: "#/components/requestBodies/InitializeStorePurchase" }
      responses:
        "200": { $ref: "#/components/responses/PurchaseSessionResponse" }
        "400": { $ref: "#/components/responses/ErrorResponse" }
  /store/purchases/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      operationId: getStorePurchase
      responses:
        "200": { $ref: "#/components/responses/PurchaseSessionResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
  /store/purchases/{id}/reconcile:
    parameters:
      - $ref: "#/components/parameters/ID"
    post:
      operationId: reconcileStorePurchase
      responses:
        "200": { $ref: "#/components/responses/PurchaseSessionResponse" }
...
  /market/preview:
    get:
      operationId: getMarketPreview
      parameters:
        - in: query
          name: marketName
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Related market item metadata.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RelatedItem" }
        "502": { $ref: "#/components/responses/ErrorResponse" }
...
  /operations/{type}:
    parameters:
      - in: path
        name: type
        required: true
        schema: { type: string }
    post:
      operationId: submitOperation
      requestBody: { $ref: "#/components/requestBodies/JsonObject" }
      responses:
        "200": { $ref: "#/components/responses/OperationReceiptResponse" }
        "400": { $ref: "#/components/responses/ErrorResponse" }
...
  /protocol-trace:
    get:
      operationId: getProtocolTrace
      parameters:
        - in: query
          name: after
          schema: { type: integer, format: uint64, minimum: 0 }
      responses:
        "200":
          description: Protocol trace entries after the requested cursor.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProtocolTraceEntry" }
...
  /storage/load:
    post:
      operationId: loadStorage
      requestBody: { $ref: "#/components/requestBodies/StorageLoad" }
      responses:
        "200": { $ref: "#/components/responses/OperationReceiptResponse" }
  /storage/move-in:
    post:
      operationId: moveIntoStorage
      requestBody: { $ref: "#/components/requestBodies/StorageMove" }
      responses:
        "200": { $ref: "#/components/responses/OperationReceiptResponse" }
  /storage/move-out:
    post:
      operationId: moveOutOfStorage
      requestBody: { $ref: "#/components/requestBodies/StorageMove" }
      responses:
        "200": { $ref: "#/components/responses/OperationReceiptResponse" }
  /containers/open:
    post:
      operationId: openContainer
      requestBody: { $ref: "#/components/requestBodies/OpenContainer" }
      responses:
        "200": { $ref: "#/components/responses/OperationReceiptResponse" }
  /tradeups/preview:
    post:
      operationId: previewTradeUp
      requestBody: { $ref: "#/components/requestBodies/JsonObject" }
      responses:
        "200": { $ref: "#/components/responses/OperationReceiptResponse" }
  /tradeups/execute:
    post:
      operationId: executeTradeUp
      requestBody: { $ref: "#/components/requestBodies/JsonObject" }
      responses:
        "200": { $ref: "#/components/responses/OperationReceiptResponse" }
  /nametags/apply:
    post:
      operationId: applyNameTag
      requestBody: { $ref: "#/components/requestBodies/SetItemName" }
      responses:
        "200": { $ref: "#/components/responses/OperationReceiptResponse" }
  /nametags/remove:
    post:
      operationId: removeNameTag
      requestBody: { $ref: "#/components/requestBodies/RemoveItemName" }
      responses:
        "200": { $ref: "#/components/responses/OperationReceiptResponse" }
  /items/delete:
    post:
      operationId: deleteItem
      requestBody: { $ref: "#/components/requestBodies/DeleteItem" }
      responses:
        "200": { $ref: "#/components/responses/OperationReceiptResponse" }
  /stattrak/swap:
    post:
      operationId: applyStatTrakSwap
      requestBody: { $ref: "#/components/requestBodies/ApplyStatTrakSwap" }
      responses:
        "200": { $ref: "#/components/responses/OperationReceiptResponse" }
  /strange-parts/apply:
    post:
      operationId: applyStrangePart
      requestBody: { $ref: "#/components/requestBodies/ApplyStrangePart" }
      responses:
        "200": { $ref: "#/components/responses/OperationReceiptResponse" }
  /items/use:
    post:
      operationId: useItem
      requestBody: { $ref: "#/components/requestBodies/UseItem" }
      responses:
        "200": { $ref: "#/components/responses/OperationReceiptResponse" }
  /items/use-multiple:
    post:
      operationId: useMultipleItems
      requestBody: { $ref: "#/components/requestBodies/UseMultipleItems" }
      responses:
        "200": { $ref: "#/components/responses/OperationReceiptResponse" }
  /tools/apply:
    post:
      operationId: applyToolToItem
      requestBody: { $ref: "#/components/requestBodies/ApplyToolToItem" }
      responses:
        "200": { $ref: "#/components/responses/OperationReceiptResponse" }
  /tools/apply-base:
    post:
      operationId: applyToolToBaseItem
      requestBody: { $ref: "#/components/requestBodies/ApplyToolToBaseItem" }
      responses:
        "200": { $ref: "#/components/responses/OperationReceiptResponse" }
  /gifts/send:
    post:
      operationId: sendGift
      requestBody: { $ref: "#/components/requestBodies/GiftItem" }
      responses:
        "200": { $ref: "#/components/responses/OperationReceiptResponse" }
components:
  parameters:
    Game:
      in: path
      name: game
      required: true
      schema: { type: string }
    AppID:
      in: path
      name: appID
      required: true
      schema: { type: integer, format: uint32, minimum: 1 }
    ID:
      in: path
      name: id
      required: true
      schema: { type: string }
  requestBodies:
    JsonObject:
      required: false
      content:
        application/json:
          schema: { $ref: "#/components/schemas/JsonObject" }
    InitializeStorePurchase:
      required: true
      content:
        application/json:
          schema:
            { $ref: "#/components/schemas/InitializeStorePurchaseRequest" }
...
    ArmoryRedeem:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ArmoryRedeemRequest" }
    OpenContainer:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/OpenContainerRequest" }
    SetItemName:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/SetItemNameRequest" }
    RemoveItemName:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ItemIDRequest" }
    DeleteItem:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ItemIDRequest" }
    ApplyStatTrakSwap:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApplyStatTrakSwapRequest" }
    ApplyStrangePart:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApplyStrangePartRequest" }
    UseItem:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/UseItemRequest" }
    UseMultipleItems:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/UseMultipleItemsRequest" }
    ApplyToolToItem:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApplyToolToItemRequest" }
    ApplyToolToBaseItem:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApplyToolToBaseItemRequest" }
    GiftItem:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/GiftItemRequest" }
...
    StorageLoad:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/StorageLoadRequest" }
    StorageMove:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/StorageMoveRequest" }
  responses:
...
    OperationReceiptResponse:
      description: Operation receipt.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/OperationReceipt" }
...
    PurchaseSessionResponse:
      description: Store purchase session.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/PurchaseSession" }
...
    GameInventoryResponse:
      description: Economy inventory snapshot.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/GameInventorySnapshot" }
    ErrorResponse:
      description: Request failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    JsonObject:
      type: object
      additionalProperties: true
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: { type: string }
...
    OperationState:
      type: string
      enum:
        - queued
        - validating
        - encoded
        - sent
        - awaiting_gc_confirmation
        - reconciling_inventory
        - completed
        - failed
        - blocked_by_feature_flag
        - requires_validation
        - requires_connection
    OperationReceipt:
      type: object
      additionalProperties: false
      required:
        - operationId
        - type
        - state
        - createdAt
      properties:
        operationId:
          type: string
        type:
          type: string
        state:
          $ref: "#/components/schemas/OperationState"
        createdAt:
          type: string
          format: date-time
        message:
          type: string
        result:
          $ref: "#/components/schemas/OperationResult"
    OperationResult:
      type: object
      additionalProperties: true
      properties:
        kind:
          type: string
          enum: [inventory_award, terminal_unsealed, terminal_offer]
        openedItem: { $ref: "#/components/schemas/InventoryItem" }
        terminalItemId: { type: string }
        terminalDefIndex: { type: integer, format: uint32, minimum: 0 }
        offerItemId: { type: string }
        offer: { $ref: "#/components/schemas/RelatedItem" }
        pointsRemaining: { type: integer, format: uint32, minimum: 0 }
        terminalOffer: { $ref: "#/components/schemas/TerminalOffer" }
        consumedItemId: { type: string }
        requestEMsg: { type: integer, format: uint32, minimum: 0 }
        requestMethod: { type: string }
        requestBodyHex: { type: string }
        confirmation: { type: string }
        responseEMsg: { type: integer, format: uint32, minimum: 0 }
        responseBodyHex: { type: string }
        beforeItemCount: { type: integer, minimum: 0 }
        afterItemCount: { type: integer, minimum: 0 }
        diagnostics:
          type: array
          items: { type: string }
...
    ProtocolTraceEntry:
      type: object
      additionalProperties: false
      required:
        [
          id,
          timestamp,
          direction,
          layer,
          emsg,
          name,
          protobuf,
          bodyBytes,
          bodyHex,
        ]
      properties:
        id: { type: integer, minimum: 0 }
        timestamp: { type: string, format: date-time }
        direction: { type: string, enum: [sent, received] }
        layer: { type: string }
        appId: { type: integer, format: uint32, minimum: 0 }
        emsg: { type: integer, format: uint32, minimum: 0 }
        name: { type: string }
        protobuf: { type: boolean }
        bodyBytes: { type: integer, minimum: 0 }
        bodyHex: { type: string }
        decoded: {}
        decodeError: { type: string }
...
    RelatedItem:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        defindex:
          type: integer
          format: uint32
          minimum: 0
        paintKit:
          type: integer
          format: uint32
          minimum: 0
        name:
          type: string
        marketName:
          type: string
        listingName:
          type: string
        kind:
          type: string
        rarity:
          type: string
        imageUrl:
          type: string
        inspectUrl:
          type: string
        price:
          type: string
        paintWear:
          type: number
          format: double
        wearMin:
          type: number
          format: double
        wearMax:
          type: number
          format: double
        items:
          type: array
          items:
            $ref: "#/components/schemas/RelatedItem"
    TerminalOffer:
      type: object
      additionalProperties: false
      required:
        - fauxItemId
        - item
      properties:
        fauxItemId:
          type: string
        generationTime:
          type: integer
          format: uint32
          minimum: 0
        purchasePrice:
          type: integer
          format: uint32
          minimum: 0
        item:
          $ref: "#/components/schemas/RelatedItem"
...
    InitializeStorePurchaseRequest:
      type: object
      additionalProperties: false
      required:
        [offerId, quantity, expectedPriceSheetVersion, expectedAmountMinor]
      properties:
        offerId: { type: string, minLength: 1 }
        quantity: { type: integer, minimum: 1, maximum: 100 }
        expectedPriceSheetVersion: { type: integer, minimum: 0 }
        expectedAmountMinor: { type: integer, minimum: 0 }
        supplementalData: { type: string }
        expectedTerminalOfferItemId: { type: string, pattern: '^[0-9]+$' }
    PurchaseSession:
      type: object
      additionalProperties: false
      required:
        [
          id,
          status,
          offerId,
          defIndex,
          name,
          quantity,
          currency,
          amountMinor,
          formattedAmount,
          createdAt,
        ]
      properties:
        id: { type: string }
        status:
          type: string
          enum:
            [
              initializing,
              awaiting_steam_authorization,
              awaiting_user,
              finalizing,
              completed,
              cancelled,
              failed,
              expired,
            ]
        offerId: { type: string }
        defIndex: { type: integer, format: uint32, minimum: 0 }
        name: { type: string }
        quantity: { type: integer, format: uint32, minimum: 1 }
        currency: { type: string }
        amountMinor: { type: integer, minimum: 0 }
        formattedAmount: { type: string }
        transactionId: { type: string }
        orderId: { type: string }
        checkoutUrl: { type: string }
        purchasedItemIds:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        message: { type: string }
        diagnostics:
          type: array
          items: { type: string }
        errorCode: { type: string }
        errorResult: { type: integer, format: int32 }
...
    ArmoryRedeemRequest:
      type: object
      additionalProperties: false
      required:
        [campaignId, redeemId, expectedCost, redeemableBalance, generationTime]
      properties:
        campaignId: { type: integer, format: uint32, minimum: 0 }
        redeemId: { type: integer, format: uint32, minimum: 0 }
        expectedCost: { type: integer, format: uint32, minimum: 0 }
        redeemableBalance: { type: integer, format: uint32, minimum: 0 }
        generationTime: { type: integer, format: uint32, minimum: 0 }
        quantity: { type: integer, format: uint32, minimum: 1 }
    ItemIDRequest:
      type: object
      additionalProperties: false
      required: [itemId]
      properties:
        itemId: { type: string, pattern: '^[0-9]+$' }
    SetItemNameRequest:
      type: object
      additionalProperties: false
      required: [subjectItemId, toolItemId, name]
      properties:
        subjectItemId: { type: string, pattern: '^[0-9]+$' }
        toolItemId: { type: string, pattern: '^[0-9]+$' }
        name: { type: string }
    OpenContainerRequest:
      type: object
      additionalProperties: false
      required: [itemId]
      properties:
        itemId: { type: string, pattern: '^[0-9]+$' }
        keyItemId: { type: string, pattern: '^[0-9]+$' }
        pointsRemaining: { type: integer, format: uint32, minimum: 0 }
        volatileLimit: { type: integer, format: uint32, minimum: 0 }
    ApplyStatTrakSwapRequest:
      type: object
      additionalProperties: false
      required: [toolItemId, item1ItemId, item2ItemId]
      properties:
        toolItemId: { type: string, pattern: '^[0-9]+$' }
        item1ItemId: { type: string, pattern: '^[0-9]+$' }
        item2ItemId: { type: string, pattern: '^[0-9]+$' }
    ApplyStrangePartRequest:
      type: object
      additionalProperties: false
      required: [strangePartItemId, itemItemId]
      properties:
        strangePartItemId: { type: string, pattern: '^[0-9]+$' }
        itemItemId: { type: string, pattern: '^[0-9]+$' }
    UseItemRequest:
      type: object
      additionalProperties: false
      required: [itemId]
      properties:
        itemId: { type: string, pattern: '^[0-9]+$' }
        targetSteamId: { type: string }
        giftPotentialTargets:
          type: array
          items: { type: integer, format: uint32, minimum: 0 }
        duelClassLock: { type: integer, format: uint32, minimum: 0 }
        initiatorSteamId: { type: string }
    UseMultipleItemsRequest:
      type: object
      additionalProperties: false
      required: [itemIds]
      properties:
        itemIds:
          type: array
          minItems: 1
          items: { type: string, pattern: '^[0-9]+$' }
    ApplyToolToItemRequest:
      type: object
      additionalProperties: false
      required: [toolItemId, subjectItemId]
      properties:
        toolItemId: { type: string, pattern: '^[0-9]+$' }
        subjectItemId: { type: string, pattern: '^[0-9]+$' }
    ApplyToolToBaseItemRequest:
      type: object
      additionalProperties: false
      required: [toolItemId, baseitemDefIndex]
      properties:
        toolItemId: { type: string, pattern: '^[0-9]+$' }
        baseitemDefIndex: { type: integer, format: uint32, minimum: 0 }
    GiftItemRequest:
      type: object
      additionalProperties: false
      required: [itemId, receiverAccountId]
      properties:
        itemId: { type: string, pattern: '^[0-9]+$' }
        receiverAccountId: { type: integer, format: uint32, minimum: 0 }
        giftMessage: { type: string }
...
    EconomyTag:
      type: object
      additionalProperties: false
      required: [category, internalName, name]
      properties:
        category: { type: string }
        internalName: { type: string }
        name: { type: string }
...
    EconomyItemDetails:
      type: object
      additionalProperties: true
      required:
        [
          game,
          level,
          qualityId,
          inventoryPosition,
          originId,
          style,
          flags,
          attributes,
        ]
      properties:
        game: { type: string, enum: [steam, steam-service, tf2, dota2] }
        level: { type: integer }
        qualityId: { type: integer }
        inventoryPosition: { type: integer }
        originId: { type: integer }
        style: { type: integer }
        flags: { type: integer }
        customName: { type: string }
        customDescription: { type: string }
        attributes:
          type: object
          additionalProperties: { type: number }
        attributeBytes:
          type: object
          additionalProperties: { type: string }
        equippedStates:
          type: array
          items: { $ref: "#/components/schemas/EquippedState" }
        interiorItemId: { type: string }
        serviceItemId: { type: string }
        serviceDefinitionId: { type: string }
        acquiredAt: { type: string }
        stateChangedAt: { type: string }
        serviceState: { type: string }
        serviceOrigin: { type: string }
        dynamicProperties:
          type: object
          additionalProperties: { type: string }
        schemaQuality: { type: string }
        equipSlot: { type: string }
        usableClasses:
          type: array
          items: { type: string }
        capabilities:
          type: object
          additionalProperties: { type: string }
        itemKind: { type: string }
        itemClass: { type: string }
        craftClass: { type: string }
        craftMaterialType: { type: string }
        toolType: { type: string }
        description: { type: string }
        collection: { type: string }
        equipRegions:
          type: array
          items: { type: string }
        schemaTags:
          type: array
          items: { type: string }
        minLevel: { type: integer }
        maxLevel: { type: integer }
        properName: { type: boolean }
        baseItem: { type: boolean }
        hidden: { type: boolean }
        rarity: { type: string }
        hero: { type: string }
        slot: { type: string }
    EconomyInventoryItem:
      type: object
      additionalProperties: false
      required:
        [
          game,
          appId,
          assetId,
          name,
          quantity,
          tradable,
          marketable,
          tags,
          details,
        ]
      properties:
        game: { type: string, enum: [steam, steam-service, tf2, dota2] }
        appId: { type: integer, format: uint32, minimum: 1 }
        contextId: { type: string }
        assetId: { type: string }
        classId: { type: string }
        instanceId: { type: string }
        definitionId: { type: integer, minimum: 0 }
        name: { type: string }
        marketName: { type: string }
        imageUrl: { type: string }
        inspectUrl: { type: string }
        quantity: { type: integer, minimum: 1 }
        type: { type: string }
        rarity: { type: string }
        quality: { type: string }
        tradable: { type: boolean }
        marketable: { type: boolean }
        tradableAfter: { type: string }
        tags:
          type: array
          items: { $ref: "#/components/schemas/EconomyTag" }
        descriptions:
          type: array
          items: { type: string }
        details: { $ref: "#/components/schemas/EconomyItemDetails" }
    GameInventorySnapshot:
      type: object
      additionalProperties: false
      required: [game, appId, items, refreshedAt, status, diagnostics]
      properties:
        game: { type: string, enum: [steam, steam-service, tf2, dota2] }
        appId: { type: integer, format: uint32, minimum: 1 }
        items:
          type: array
          items: { $ref: "#/components/schemas/EconomyInventoryItem" }
        refreshedAt: { type: string, format: date-time }
        status:
          { type: string, enum: [ready, requires_connection, loading, error] }
        message: { type: string }
        error: { type: string }
        schemaRevision: { type: string }
        diagnostics:
          type: array
          items: { type: string }
...
    StorageLoadRequest:
      type: object
      additionalProperties: false
      required: [casketId]
      properties:
        casketId: { type: string, pattern: '^[0-9]+$' }
    StorageMoveRequest:
      type: object
      additionalProperties: false
      required: [casketId, itemId]
      properties:
        casketId: { type: string, pattern: '^[0-9]+$' }
        itemId: { type: string, pattern: '^[0-9]+$' }
...